git-epic-creator/services/shared/pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "shared"
version = "0.1.0"
description = "Common shared library for requirement ai generator tool services."
authors = [
    { name = "Vladyslav Avilov", email = "vladyslav.avilov@email.com" }
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "structlog",
    "fastapi",
    "neo4j",
    "pydantic",
    "pydantic-settings", 
    "python-dotenv",
    "sqlalchemy",
    "sqlalchemy-utils",
    "fastapi-azure-auth",
    "redis[hiredis]",
    "celery"
]

[tool.pytest.ini_options]
pythonpath = ["src"]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
filterwarnings = [
    "ignore::PendingDeprecationWarning"
]

[tool.setuptools.packages.find]
where = ["src"]
exclude = ["tests*"]

[tool.setuptools]
py-modules = ["models"]

[tool.setuptools.package-dir]
"" = "src"

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-asyncio"
]